update price column from Double to DECIMAL(10, 2) to avoid precision loss

+++ sql/menu_sku.sql
@@ -2,11 +2,11 @@ create table menu_sku
 (
     id           bigint auto_increment
         primary key,
-    is_available bit         not null,
-    price        double      not null,
-    stock        int         not null,
-    menu_id      bigint      not null,
-    size         varchar(20) not null,
+    is_available bit            not null,
+    price        decimal(10, 2) not null,
+    stock        int            not null,
+    menu_id      bigint         not null,
+    size         varchar(20)    not null,
     constraint unique_menu_size
         unique (menu_id, size),
     constraint FKchetvl74h8evnny3vy8fd9j36
